+++ packages/core/src/languages/css/css-values.ts
@@ -20,14 +20,16 @@ function newCssDeclarationValueExpression(
   return new SingleGroupMangleExpression(
     `
       (?<=
-        {[^\\}]*
-        \\:\\s*
+        \\{
+        [^\\}]+
+        :\\s*
+        ([^;]*\\s)?
         ${valuePrefix}
       )
       (?<${GROUP_MAIN}>%s)
       (?=
         ${valueSuffix}
-        \\s*(?:\\;|\\})
+        (?:\\s|\\;|\\})
       )
     `,
     GROUP_MAIN,
@@ -42,7 +44,7 @@ function newCssDeclarationValueExpression(
  * @param options The {@link CssDeclarationValueOptions}.
  * @returns A set of {@link MangleExpression}s.
  * @since v0.1.14
- * @version v0.1.18
+ * @version v0.1.19
  */
 export default function cssDeclarationValueExpressionFactory(
   options: CssDeclarationValueOptions,