+++ packages/core/src/languages/html/__tests__/attributes.test.ts
@@ -4,7 +4,7 @@ import type { AttributeOptions } from "../../options";
 
 import { expect } from "chai";
 
-import { matchesAsExpected } from "../../__tests__/test-helpers";
+import { getAllMatches } from "../../__tests__/test-helpers";
 
 import attributeExpressionFactory from "../attributes";
 
@@ -14,33 +14,33 @@ suite("HTML - Attribute Expression Factory", function() {
       name: "sample",
       cases: [
         {
-          input: "<div data-foo></div>",
+          input: "<div data-foobar></div>",
           pattern: "[a-z\\-]+",
-          expected: ["data-foo"],
+          expected: ["data-foobar", "data-foobar"],
           options: null,
         },
         {
           input: "<div data-foo=\"bar\"></div>",
           pattern: "[a-z\\-]+",
-          expected: ["data-foo"],
+          expected: ["data-foo", "data-foo"],
           options: null,
         },
         {
           input: "<div data-foo=\"bar\" data-bar=\"foo\"></div>",
           pattern: "[a-z\\-]+",
-          expected: ["data-foo", "data-bar"],
+          expected: ["data-foo", "data-bar", "data-foo"],
           options: null,
         },
         {
-          input: "<div data-foo='bar'></div>",
+          input: "<div data-hello='world'></div>",
           pattern: "[a-z\\-]+",
-          expected: ["data-foo"],
+          expected: ["data-hello", "data-hello"],
           options: null,
         },
         {
-          input: "<div data-foo='bar' data-bar='foo'></div>",
+          input: "<div data-foo='baz' data-baz='foo'></div>",
           pattern: "[a-z\\-]+",
-          expected: ["data-foo", "data-bar"],
+          expected: ["data-foo", "data-foo", "data-baz"],
           options: null,
         },
       ],
@@ -57,8 +57,8 @@ suite("HTML - Attribute Expression Factory", function() {
         } = testCase;
 
         const expressions = attributeExpressionFactory();
-        const result = matchesAsExpected(expressions, input, pattern, expected);
-        expect(result).to.equal(true, `in "${input}"`);
+        const matches = getAllMatches(expressions, input, pattern);
+        expect(matches).to.deep.equal(expected);
       }
     });
   }