+++ packages/core/src/languages/html/multi-value-attributes.ts
@@ -1,7 +1,10 @@
 import type { MangleExpression } from "../../types";
 import type { MultiValueAttributeOptions } from "../options";
 
-import { NestedGroupExpression } from "../utils/mangle-expressions";
+import {
+  NestedGroupMangleExpression,
+  SingleGroupMangleExpression,
+} from "../utils/mangle-expressions";
 import { QUOTED_ATTRIBUTE_PATTERN, QUOTES_ARRAY } from "./common";
 
 const GROUP_MAIN = "main";
@@ -16,7 +19,7 @@ const GROUP_MAIN = "main";
 function newElementAttributeMultiValueExpressions(
   attributesPattern: string,
 ): MangleExpression[] {
-  return QUOTES_ARRAY.map((quote) => new NestedGroupExpression(
+  return QUOTES_ARRAY.map((quote) => new NestedGroupMangleExpression(
     `
       (?<=
         \\<\\s*[a-zA-Z0-9]+\\s+
@@ -48,6 +51,36 @@ function newElementAttributeMultiValueExpressions(
 }
 
 /**
+ * Get {@link MangleExpression}s to match unquoted element attribute values in
+ * HTML, e.g. `bar` in `<div data-foo=bar></div>`.
+ *
+ * @param attributesPattern The pattern of attribute names.
+ * @returns The {@link MangleExpression}s to match unquoted attribute values.
+ */
+function newUnquotedAttributeValueExpression(
+  attributesPattern: string,
+): MangleExpression[] {
+  return QUOTES_ARRAY.map((quote) => new SingleGroupMangleExpression(
+    `
+      (?<=
+        \\<\\s*[a-zA-Z0-9]+\\s+
+        (?:
+          [^>\\s=]+
+          (?:\\s*=\\s*${quote}[^${quote}]*${quote})?
+          \\s+
+        )*
+        (?:${attributesPattern})\\s*=\\s*
+      )
+      (?<${GROUP_MAIN}>%s)
+      (?=
+        (?:\\s|\\/|\\>)
+      )
+    `,
+    GROUP_MAIN,
+  ));
+}
+
+/**
  * Get the set of {@link MangleExpression}s to match multi-value attribute
  * values in HTML. This will match:
  * - Attribute values (e.g. `the` and `sun` in `<img data-praise="the sun">`).
@@ -55,7 +88,7 @@ function newElementAttributeMultiValueExpressions(
  * @param options The {@link MultiValueAttributeOptions}.
  * @returns A set of {@link MangleExpression}s.
  * @since v0.1.14
- * @version v0.1.18
+ * @version v0.1.19
  */
 export default function multiValueAttributeExpressionFactory(
   options: MultiValueAttributeOptions,
@@ -64,5 +97,6 @@ export default function multiValueAttributeExpressionFactory(
 
   return [
     ...newElementAttributeMultiValueExpressions(attributesPattern),
+    ...newUnquotedAttributeValueExpression(attributesPattern),
   ];
 }