+++ packages/core/src/languages/utils/mangle-expressions/index.ts
@@ -1,2 +1,3 @@
 export { default as NestedGroupExpression } from "./nested-group.class";
+export { default as NestedGroupMangleExpression } from "./nested-group.class";
 export { default as SingleGroupMangleExpression } from "./single-group.class";