+++ packages/core/src/manglers/__tests__/css-variables.test.ts
@@ -128,6 +128,10 @@ suite("CSS Variable Mangler", function() {
             input: "div { color: var(--foo); } p { font: var(--bar); }",
             expected: "div { color: var(--a); } p { font: var(--b); }",
           },
+          {
+            input: "div { margin: 0 var(--foo); }",
+            expected: "div { margin: 0 var(--a); }",
+          },
         ],
       },
       {
@@ -278,6 +282,7 @@ suite("CSS Variable Mangler", function() {
     const embedDeclarationsInStyle = embedAttributeValue("style");
 
     const varyAttributeSpacing = varySpacing("=");
+    const varyDeclarationSpacing = varySpacing([":", ",", ";"]);
     const varyTagSpacing = varySpacing(["<", ">"]);
 
     type TestInstance = {
@@ -297,15 +302,6 @@ suite("CSS Variable Mangler", function() {
         ],
       },
       {
-        name: "variable declaration (spacing)",
-        factory: (before: string, after: string): TestCase[] => [
-          ...varySpacing([":", ";"], {
-              input: `--${before}:42;`,
-              expected: `--${after}:42;`,
-          }),
-        ],
-      },
-      {
         name: "variable usage without default",
         factory: (before: string, after: string): TestCase[] => [
           ...CSS_PROPERTIES
@@ -316,36 +312,23 @@ suite("CSS Variable Mangler", function() {
         ],
       },
       {
-        name: "variable usage without default (spacing)",
-        factory: (before: string, after: string): TestCase[] => [
-          ...varySpacing(["(", ")"], {
-              input: `content:var(--${before});`,
-              expected: `content:var(--${after});`,
-          }),
-        ],
-      },
-      {
-        name: "variable usage with default",
+        name: "variable usage with default (different properties)",
         factory: (before: string, after: string): TestCase[] => [
           ...CSS_PROPERTIES
             .map((property: string): TestCase => ({
               input: `${property}:var(--${before},42);`,
               expected: `${property}:var(--${after},42);`,
             })),
+        ],
+      },
+      {
+        name: "variable usage with default (different values)",
+        factory: (before: string, after: string): TestCase[] => [
           ...CSS_VALUES_NO_STRINGS
             .map((value: string): TestCase => ({
               input: `content:var(--${before},${value});`,
               expected: `content:var(--${after},${value});`,
             })),
-        ],
-      },
-      {
-        name: "variable usage with default (spacing)",
-        factory: (before: string, after: string): TestCase[] => [
-          ...varySpacing(["(", ","], {
-              input: `content:var(--${before},42);`,
-              expected: `content:var(--${after},42);`,
-          }),
         ],
       },
     ];
@@ -362,6 +345,20 @@ suite("CSS Variable Mangler", function() {
             .filter((testCase) => !/(\s|^)(style)=/.test(testCase.input)),
         ]
         .flatMap(varyHtmlQuotes)
+        .flatMap(embedAttributesInTags),
+      },
+      {
+        name: "multi-value CSS declaration",
+        cases: [
+          {
+            input: "style=\"margin: 0 var(--foobar);\"",
+            expected: "style=\"margin: 0 var(--a);\"",
+          },
+          {
+            input: "style=\"margin: 1em var(--foo) 1px var(--bar);\"",
+            expected: "style=\"margin: 1em var(--a) 1px var(--b);\"",
+          },
+        ]
         .flatMap(embedAttributesInTags),
       },
       {
@@ -436,6 +433,30 @@ suite("CSS Variable Mangler", function() {
             .map(embedDeclarationsInStyle)
             .flatMap(varyAttributeSpacing)
             .flatMap(varyHtmlQuotes)
+            .flatMap(embedAttributesInTags),
+        },
+        {
+          name: `${name}, vary spacing`,
+          cases: factory("foobar", "a")
+            .map(embedDeclarationsInStyle)
+            .flatMap(varyDeclarationSpacing)
+            .flatMap(embedAttributesInTags),
+        },
+        {
+          name: `${name} in an unquoted style attribute`,
+          cases: factory("foobar", "a")
+            .map(embedDeclarationsInStyle)
+            .flatMap((testCase: TestCase): TestCase[] => [
+              {
+                input: testCase.input.replace(/"/g, ""),
+                expected: testCase.expected.replace(/"/g, ""),
+              },
+              {
+                input: testCase.input.replace(/("|;)/g, ""),
+                expected: testCase.expected.replace(/("|;)/g, ""),
+              },
+            ])
+            .flatMap(varyAttributeSpacing)
             .flatMap(embedAttributesInTags),
         },
         {