+++ packages/benchmarking/CHANGELOG.md
@@ -10,6 +10,10 @@ Versioning].
 
 - _No changes yet_
 
+## [0.1.1] - 2021-04-30
+
+- Add option to setup individual benchmark runs.
+
 ## [0.1.0] - 2021-04-24
 
 - Benchmark a function using `benchmarkFn`.